initial conversion of occupancies from integer to double

+++ src/qmfunctions/orbital_utils.h
@@ -79,17 +79,17 @@ int size_paired(const OrbitalVector &Phi);
 int size_alpha(const OrbitalVector &Phi);
 int size_beta(const OrbitalVector &Phi);
 int get_multiplicity(const OrbitalVector &Phi);
-int get_electron_number(const OrbitalVector &Phi, int spin = SPIN::Paired);
+double get_electron_number(const OrbitalVector &Phi, int spin = SPIN::Paired);
 int start_index(const OrbitalVector &Phi, int spin);
 int get_n_nodes(const OrbitalVector &Phi, bool avg = false);
 int get_size_nodes(const OrbitalVector &Phi, bool avg = false);
 bool orbital_vector_is_sane(const OrbitalVector &Phi);
 
 void set_spins(OrbitalVector &Phi, const IntVector &spins);
-void set_occupations(OrbitalVector &Phi, const IntVector &occ);
+void set_occupations(OrbitalVector &Phi, const DoubleVector &occ);
 
 IntVector get_spins(const OrbitalVector &Phi);
-IntVector get_occupations(const OrbitalVector &Phi);
+DoubleVector get_occupations(const OrbitalVector &Phi);
 DoubleVector get_norms(const OrbitalVector &Phi);
 DoubleVector get_squared_norms(const OrbitalVector &Phi);
 DoubleVector calc_eigenvalues(const OrbitalVector &Phi, const ComplexMatrix &F_mat);
